deps/v8/test/inspector/debugger/evaluate-with-await-on-breakpoint-expected.txt

Test evaluating await expression on a breakpoint

Running test: testScopesPaused
Evaluating await expression
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            className : SyntaxError
            description : SyntaxError: await can not be used when evaluating code while paused in the debugger     at run (<anonymous>:3:3)     at <anonymous>:1:1
            objectId : 1.1.7
            subtype : error
            type : object
        }
        exceptionId : 1
        lineNumber : 0
        scriptId : 5
        text : Uncaught
    }
    result : {
        className : SyntaxError
        description : SyntaxError: await can not be used when evaluating code while paused in the debugger     at run (<anonymous>:3:3)     at <anonymous>:1:1
        objectId : 1.1.6
        subtype : error
        type : object
    }
}
Evaluating await expression in async function
{
    result : {
        className : Promise
        description : Promise
        objectId : 1.1.8
        subtype : promise
        type : object
    }
}
